Backend/langchain-backend/requirements.txt
langchain
langsmith
openai
python-dotenv